pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.4",
    "pyyaml>=6.0.3",
    "scipy>=1.16.3",
    "tqdm>=4.67.1",
]
